+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.5] - 2023-08-24
+
+### Added
+
+- Added the [player](https://developer.spotify.com/documentation/web-api/reference/get-information-about-the-users-current-playback) endpoints.
+Two of them return `403 Forbidden "Player command failed: Restriction violated"` (set repeat mode and toggle playback shuffle).
+It seems to be an issue with the Spotify API again.
+
 ## [0.3.4] - 2023-08-21
 
 ### Added